Move common check of Common Name in certificate and name used in json payload or path to a common security module
