+++ neighbour-server/src/test/java/no/vegvesen/ixn/federation/server/NeighbourRestControllerTest.java
@@ -1,6 +1,7 @@
 package no.vegvesen.ixn.federation.server;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import no.vegvesen.ixn.federation.auth.CertService;
 import no.vegvesen.ixn.federation.api.v1_0.*;
 import no.vegvesen.ixn.federation.exceptions.InterchangeNotInDNSException;
 import no.vegvesen.ixn.federation.exceptions.SubscriptionRequestException;
@@ -11,6 +12,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.context.annotation.Import;
 import org.springframework.http.MediaType;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContext;
@@ -28,6 +30,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @WebMvcTest
+@Import(CertService.class)
 class NeighbourRestControllerTest {
 
 	private MockMvc mockMvc;
@@ -40,7 +43,6 @@ class NeighbourRestControllerTest {
 
 	private String subscriptionRequestPath = "/subscription";
 	private String capabilityExchangePath = "/capabilities";
-
 
 	private Set<String> quadTree = Collections.emptySet();
